+++ Sources/KeyboardKit/System/SystemKeyboardButton.swift
@@ -2,92 +2,154 @@
 //  SystemKeyboardButton.swift
 //  KeyboardKit
 //
-//  Created by Daniel Saidi on 2021-10-05.
-//  Copyright © 2021 Daniel Saidi. All rights reserved.
+//  Created by Daniel Saidi on 2020-07-02.
+//  Copyright © 2020-2023 Daniel Saidi. All rights reserved.
 //
-
-import Foundation
 
 import SwiftUI
 
 /**
- This view can be used to create a keyboard button that aims
- to mimic a native iOS keyboard button.
+ This view renders a system keyboard button that is based on
+ a certain ``KeyboardAction``.
 
- The view wraps any content and apply a certain button style
- to it, which by default is the style of a native iOS button.
- 
- Note that this view does not add any gestures or actions to
- the button. It only renders the button content. You can use
- a ``SystemKeyboardActionButton`` to get a functional button
- with gestures, pressed state etc.
+ This view will adapt its content to conform to the provided
+ `action` and `appearance` and applies keyboard gestures for
+ the provided `action`, `actionHandler` and `context`.
+
+ The view sets up gestures, line limits, vertical offset etc.
+ and aims to make the button mimic an iOS system keyboard as
+ closely as possible. You can however use the `contentConfig`
+ parameter to customize or replace the content view.
  */
 public struct SystemKeyboardButton<Content: View>: View {
-    
+
     /**
-     Create a system keyboard button.
-     
+     Create a system keyboard button view.
+
      - Parameters:
-       - content: The button content.
-       - style: The style to apply to the button
+       - action: The keyboard action to apply.
+       - actionHandler: The action handler to use.
+       - keyboardContext: The keyboard context to which the button should apply.
+       - calloutContext: The callout context to affect, if any.
+       - appearance: The appearance to apply to the button.
+       - contentConfig: A configuration block that can be used to customize or replace the standard button content.
      */
     public init(
-        content: Content,
-        style: KeyboardButtonStyle
+        action: KeyboardAction,
+        actionHandler: KeyboardActionHandler,
+        keyboardContext: KeyboardContext,
+        calloutContext: KeyboardCalloutContext?,
+        appearance: KeyboardAppearance,
+        contentConfig: @escaping ContentConfig
     ) {
-        self.content = content
-        self.style = style
+        self.action = action
+        self.actionHandler = actionHandler
+        self.keyboardContext = keyboardContext
+        self.calloutContext = calloutContext
+        self.appearance = appearance
+        self.contentConfig = contentConfig
+    }
+
+    /**
+     Create a system keyboard button view.
+
+     - Parameters:
+       - action: The keyboard action to apply.
+       - actionHandler: The action handler to use.
+       - keyboardContext: The keyboard context to which the button should apply.
+       - calloutContext: The callout context to affect, if any.
+       - appearance: The appearance to apply to the button.
+     */
+    public init(
+        action: KeyboardAction,
+        actionHandler: KeyboardActionHandler,
+        keyboardContext: KeyboardContext,
+        calloutContext: KeyboardCalloutContext?,
+        appearance: KeyboardAppearance
+    ) where Content == SystemKeyboardButtonContent {
+        self.init(
+            action: action,
+            actionHandler: actionHandler,
+            keyboardContext: keyboardContext,
+            calloutContext: calloutContext,
+            appearance: appearance,
+            contentConfig: { $0 }
+        )
     }
     
-    private let content: Content
-    private let style: KeyboardButtonStyle
+    private let action: KeyboardAction
+    private let actionHandler: KeyboardActionHandler
+    private let keyboardContext: KeyboardContext
+    private let calloutContext: KeyboardCalloutContext?
+    private let appearance: KeyboardAppearance
+    private let contentConfig: ContentConfig
+
+    @State
+    private var isPressed = false
     
+    /**
+     This typealias represents an action that can be used to
+     customize (or replace) a standard button content view.
+     */
+    public typealias ContentConfig = (_ standardContent: SystemKeyboardButtonContent) -> Content
+        
     public var body: some View {
-        content
-            .systemKeyboardButtonStyle(style)
+        buttonContent
+            .systemKeyboardButtonStyle(buttonStyle)
+            .keyboardGestures(
+                for: action,
+                actionHandler: actionHandler,
+                calloutContext: calloutContext,
+                isPressed: $isPressed
+            )
+    }
+}
+
+private extension SystemKeyboardButton {
+    
+    var buttonContent: some View {
+        contentConfig(
+            SystemKeyboardButtonContent(
+                action: action,
+                appearance: appearance,
+                keyboardContext: keyboardContext
+            )
+        )
+    }
+    
+    var buttonStyle: KeyboardButtonStyle {
+        appearance.buttonStyle(
+            for: action,
+            isPressed: isPressed
+        )
     }
 }
 
 struct SystemKeyboardButton_Previews: PreviewProvider {
     
-    static let appearance = StandardKeyboardAppearance(
-        keyboardContext: .preview)
-    
-    static func previewButton<Content: View>(for content: Content, style: KeyboardButtonStyle) -> some View {
+    static func button(for action: KeyboardAction) -> some View {
         SystemKeyboardButton(
-            content: content
-                .padding(.horizontal, 80)
-                .padding(.vertical, 20),
-            style: style)
-    }
-    
-    static var previewImage: some View {
-        Image("photo-forest", bundle: .keyboardKit)
-            .resizable()
-            .aspectRatio(contentMode: .fill)
-            .edgesIgnoringSafeArea(.all)
-    }
-    
-    static var previewStack: some View {
-        ScrollView {
-            VStack(spacing: 20) {
-                previewButton(for: Text("hello"), style: .preview1)
-                previewButton(for: Text("HELLO"), style: .preview2)
-                previewButton(for: SystemKeyboardButtonText(text: "🚀", action: .character("")), style: .preview1)
-                previewButton(for: Image.keyboardGlobe, style: .preview2)
-                
-                previewButton(for: Text("input"), style: appearance.buttonStyle(for: .character(""), isPressed: false))
-                previewButton(for: Text("input pressed"), style: appearance.buttonStyle(for: .character(""), isPressed: true))
-                previewButton(for: Text("control"), style: appearance.buttonStyle(for: .backspace, isPressed: false))
-                previewButton(for: Text("control pressed"), style: appearance.buttonStyle(for: .backspace, isPressed: true))
-            }.padding(.top, 40)
+            action: action,
+            actionHandler: .preview,
+            keyboardContext: .preview,
+            calloutContext: .preview,
+            appearance: .preview
+        ) {
+            $0.frame(width: 80, height: 80)
         }
     }
     
     static var previews: some View {
-        ZStack {
-            previewImage
-            previewStack
+        VStack {
+            button(for: .backspace)
+            button(for: .space)
+            button(for: .nextKeyboard)
+            button(for: .character("a"))
+            button(for: .character("A"))
         }
+        .padding()
+        .background(Color.gray)
+        .cornerRadius(10)
+        .environment(\.sizeCategory, .extraExtraLarge)
     }
 }